chore: Update the precision of the updated_at timestamp in conversation model (#10875)
Use to_f instead of to_i to preserve the millisecond precision in the UI.

+++ app/presenters/conversations/event_data_presenter.rb
@@ -43,7 +43,7 @@ class Conversations::EventDataPresenter < SimpleDelegator
       last_activity_at: last_activity_at.to_i,
       timestamp: last_activity_at.to_i,
       created_at: created_at.to_i,
-      updated_at: updated_at.to_i
+      updated_at: updated_at.to_f
     }
   end
 end
